executors/transcribe_model/requirements.txt
openai-whisper==20231117
openai==1.16.2
docarray[audio]